+++ cymetric/__init__.py
@@ -13,7 +13,7 @@ try:
     from cymetric.evaluator import METRIC_REGISTRY, register_metric, \
         raw_to_series, Evaluator, eval
     from cymetric.execution import ExecutionContext, exec_code
-    from cymetric.eco_inputs import swu
+    from cymetric.eco_tools import swu
     from cymetric.eco_metrics import capital_cost
     from cymetric.cash_flows import lcoe
 except ImportError:
@@ -29,7 +29,7 @@ except ImportError:
     from .evaluator import METRIC_REGISTRY, register_metric, \
         raw_to_series, Evaluator, eval
     from .execution import ExecutionContext, exec_code
-    from .eco_inputs import swu
+    from .eco_tools import swu
     from .eco_metrics import capital_cost
     from .cash_flows import lcoe
